refactor(billet): store montantTtc as euros (Float) instead of centimes (Int)
Simpler for the user — enter 450 for 450 EUR, no mental conversion.
Migration changes Int to Float on billet.montantTtc and paiement.montantTtc.
Removed all /100 and *100 conversions in actions, schemas, and display.

+++ lib/schemas/billet.ts
@@ -23,7 +23,7 @@ export const billetCreateSchema = z.object({
   payeurCp: z.string().optional().or(z.literal('')),
   payeurVille: z.string().optional().or(z.literal('')),
 
-  montantTtc: z.coerce.number().int().nonnegative('Montant invalide'),
+  montantTtc: z.coerce.number().nonnegative('Montant invalide'),
   categorie: z.string().optional().or(z.literal('')),
   provenance: z.string().optional().or(z.literal('')),
   lieuDecollage: z.string().optional().or(z.literal('')),
